moves actors package to node internals
